Move bannerSubtitle under banner folder
