+++ src/components/banner/bannerSubtitle/_spec.js
