+++ src/components/banner/bannerSubtitle/index.js
