+++ src/components/bannerImage/_spec.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { shallow, mount } from 'enzyme'
 import toJSON from 'enzyme-to-json'
 import BannerTitle from '../bannerTitle'
-import BannerSubtitle from '../bannerSubtitle'
+import BannerSubtitle from '../banner/bannerSubtitle'
 import BannerImage from '.'
 import 'jest-styled-components'
